moved convective cfl calculator outside of shallow_water utils
