+++ utils/__init__.py
@@ -2,3 +2,4 @@ import utils.units  # noqa: F401
 import utils.planets  # noqa: F401
 import utils.shallow_water  # noqa: F401
 import utils.mg  # noqa: F401
+import utils.diagnostics  # noqa: F401
